requirements.txt
datetime==4.4
hydralit_components==1.0.9
numpy==1.22.2
pandas==1.4.0
plotly.express==0.4.1
plotly==5.6.0
requests==2.27.1
streamlit==1.5.1
Pillow==9.0.1
streamlit-lottie==0.0.3
streamlit-aggrid==0.2.3.post2
openpyxl==3.0.9